Split DeckGL.react.js into hooks (fixes #26)
Component drops from 924 to 473 lines (~230 excluding the propTypes
block, which must stay for react-docgen/dash-generate-components).
Behavior-preserving extraction into src/lib/components/hooks/:

- useDrawing: feature-collection sync, selection, deleteSelected
  handling, cursor, editable layer composition
- useTimeFilterAnimation: rAF engine, head-time seeding, filterRange
  application to whichever renderer is live
- useMaplibre: map + MapboxOverlay lifecycle, style switching, custom
  sources/mapLayers, moveend reporting, interaction toggles, overlay
  layer updates, controlled view sync, native fitBounds
- useDeckEvents: view-state/click/hover handlers + tooltip renderer

Shared renderer refs (overlayRef/deckRef) are created in the component
and passed to both hooks, resolving the animation<->maplibre coupling.
Debug logging moves to utils/debug.js. eslint-plugin-react-hooks added
with rules-of-hooks: error - the class of bug fixed in #9 is now
caught statically (and this refactor passes it).

Prerequisites #9 and #10 are merged into this branch per the issue's
ordering note.

Verified: all 7 browser smoke tests (drawing, time filter, interleaved,
basemap diff-swap, maplibre toggle, click events, @@scale), 53 vitest
unit tests, 239 Python tests, eslint clean with zero suppressions,
build:py output byte-identical (no generated API drift).
