+++ package.json
@@ -49,6 +49,7 @@
     "eslint-config-prettier": "^10.0.0",
     "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-react": "^7.37.0",
+    "eslint-plugin-react-hooks": "^7.1.1",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-docgen": "^5.4.3",
